backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.3.6
redis==5.0.4
google-generativeai==0.7.2
requests==2.32.3
pillow==10.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
aiofiles==23.2.1
